+++ .prettierrc
@@ -2,6 +2,6 @@
   "endOfLine": "lf",
   "semi": false,
   "singleQuote": false,
-  "tabWidth": 2,
+  "tabWidth": 4,
   "trailingComma": "es5"
 }
