--- functions-src/token-hider/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "token-hider",
-  "version": "1.0.0",
-  "description": "netlify functions:create - how to hide API tokens from your users",
-  "main": "token-hider.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [
-    "netlify",
-    "serverless",
-    "apis",
-    "js"
-  ],
-  "author": "Netlify",
-  "license": "MIT",
-  "dependencies": {
-    "axios": "^0.19.2",
-    "qs": "^6.7.0"
-  }
-}
