+++ gatsby-browser.js
@@ -1,6 +1,4 @@
 // custom typefaces
 import "typeface-montserrat"
 import "typeface-merriweather"
-
-import "prismjs/themes/prism.css"
 import "./src/styles/global.css"
