+++ package.json
@@ -5,8 +5,10 @@
   "version": "0.1.0",
   "author": "Nicholas Keeney <nkeeney@oceanics.io>",
   "dependencies": {
+    "@babel/core": "^7.10.4",
     "@babel/plugin-syntax-import-meta": "^7.10.1",
-    "@reach/dialog": "^0.2.9",
+    "@reach/dialog": "^0.10.5",
+    "@reach/portal": "^0.10.5",
     "@reach/router": "^1.3.3",
     "@reach/tabs": "^0.1.6",
     "@reach/visually-hidden": "^0.1.4",
@@ -14,23 +16,18 @@
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-preset-gatsby": "^0.4.9",
     "gatsby": "^2.22.9",
-    "gatsby-cli": "^2.12.46",
+    "gatsby-cli": "^2.12.58",
     "gatsby-image": "^2.2.39",
     "gatsby-link": "^2.2.30",
     "gatsby-plugin-create-client-paths": "^2.1.6",
-    "gatsby-plugin-feed": "^2.3.26",
-    "gatsby-plugin-google-analytics": "^2.1.34",
     "gatsby-plugin-manifest": "^2.2.39",
-    "gatsby-plugin-netlify-identity": "^0.0.3",
     "gatsby-plugin-offline": "^3.0.32",
     "gatsby-plugin-react-helmet": "^3.1.21",
     "gatsby-plugin-sharp": "^2.4.3",
-    "gatsby-plugin-typography": "^2.3.21",
     "gatsby-react-router-scroll": "^2.1.23",
     "gatsby-remark-copy-linked-files": "^2.1.36",
     "gatsby-remark-images": "^3.1.42",
     "gatsby-remark-katex": "^3.1.23",
-    "gatsby-remark-prismjs": "^3.3.30",
     "gatsby-remark-responsive-iframe": "^2.2.31",
     "gatsby-remark-smartypants": "^2.1.20",
     "gatsby-source-filesystem": "^2.1.46",
@@ -38,13 +35,11 @@
     "gatsby-transformer-sharp": "^2.5.3",
     "katex": "^0.11.1",
     "mapbox-gl": "^1.5.0",
-    "neo4j-driver": "^4.0.2",
-    "neo4j-graphql-js": "^2.14.2",
-    "prismjs": "^1.19.0",
+    "prismjs": "^1.20.0",
     "prop-types": "^15.7.2",
-    "react-dom": "^16.12.0",
+    "react-dom": "^16.13.1",
     "react-helmet": "^5.2.1",
-    "react-netlify-identity-widget": "^0.2.7",
+    "react-router-dom": "^5.2.0",
     "react-typography": "^0.16.19",
     "styled-components": "^5.1.1",
     "typeface-merriweather": "0.0.72",
@@ -53,6 +48,7 @@
     "typography-theme-wordpress-2016": "^0.16.19"
   },
   "devDependencies": {
+    "eslint": "^7.4.0",
     "prettier": "^1.19.1"
   },
   "homepage": "https://www.oceanics.io",
@@ -66,10 +62,8 @@
     "url": "git+https://bitbucket.org/oceanicsdotio/oceanics.io.git"
   },
   "scripts": {
-    "build-wasm": "wasm-pack build --no-typescript --out-dir src/wasm",
     "build": "gatsby build",
     "develop": "gatsby develop",
-    "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
     "start": "npm run develop",
     "serve": "gatsby serve",
     "clean": "gatsby clean",
