+++ src/components/Table.js
@@ -1,8 +1,8 @@
-import React from "react";
+import React, {useState, useEffect} from "react";
 import styled from "styled-components";
 
 const StyledCell = styled.td`
-    padding: 3px;
+    padding: 1px;
     margin: 0;
 `;
 
@@ -10,21 +10,38 @@ const StyledInput = styled.input`
     position: relative;
     width: 100%;
     height: 100%;
-    padding: 2px;;
+    padding: 1px;;
     margin: 1px;
     text-align: left;
     border: solid 1px #333333;
-    border-radius: 5px;
+    border-radius: 3px;
     background: #202020;
     font-family: inherit;
-    font-size: smaller;
     color: #CCCCCC;
     text-decoration: none;
+    overflow: hidden;
 
-    &:hover {
-        border: solid 1px orange;
+    &:focus {
+        border-color: #77CCFF;
     }
 
+    &:hover:not(:focus) {
+
+        border-color: #EF5FA1;
+        animation: scroll-left 5s linear 1;
+    
+        @keyframes scroll-left {
+            0%   { 
+                text-indent: 0; 		
+            }
+            100% {
+                text-indent: -50%; 
+            }
+        }
+    }
+
+    
+    
 `;
 
 const StyledRow = styled.tr`
@@ -45,25 +62,14 @@ const StyledTable = styled.table`
 
 const StyledHead = styled.th`
     padding: 0px;
+    color: #77CCFF;
     /* background: #202020; */
     /* border: 1px solid black; */
 `;
 
-const StyledFoot = styled.tfoot``;
+const EditableCell = ({ record, col: { label, format, parse }, ind }) => {
 
-const evt = () => {
-    console.log("onBlur trigger");
-}
-
-const EditableCell = (props) => {
-
-    const { record, col, ind } = props;
-    const { label, format, parse } = col;
-
-    
     let value = record[label];
-
-  
     if (typeof value === typeof "") {
         value = value.trim();
     } else if (typeof value === "object" && value !== null) {
@@ -72,63 +78,74 @@ const EditableCell = (props) => {
     return (
         <StyledCell key={ind}>
             <StyledInput 
-                onBlur={({target}) => record[label] = parse ? parse(target.value) : target.value } 
+                onBlur={({target}) => {
+                    record[label] = parse ? parse(target.value) : target.value;
+                }} 
                 defaultValue={format ? format(value) : value} />
         </StyledCell>
     )
 };
 
-export const RecordRow = (props) => {
-
-    const { schema, record, ind } = props;
-
+export const RecordRow = ({ schema, record, ind }) => {
     return (
         <StyledRow key={ind}>
-            {/* <StyledHead key={ind} scope={"row"}>{ind}</StyledHead> */}
             {schema.map((key, ii) => <EditableCell record={record} col={key} ind={ii} />)}
         </StyledRow>
     )
 };
 
+
+const CallbackHeader = ({label, key, onClick=null}) => 
+    <StyledHead key={key} scope={"col"} onClick={onClick}>
+        {label.replace(/([a-z](?=[A-Z]))/g, '$1 ').toLowerCase()}
+    </StyledHead>;
+
+
 export default (props) => {
 
-    const { records, order } = props;
-    let { schema } = props;
+    const [schema, setSchema] = useState(props.schema)
+    const [records, setRecords] = useState(props.records)
+ 
+    useEffect(()=>{
+
+        if (schema === undefined) {
+            const _implicitSchema = records.map(e => {
+                return new Set(Object.keys(e).filter(key => !key.includes("@")))
+            }).reduce(
+                (acc, current) => new Set([...acc, ...current])
+            );
     
+            let priority = [];
+            ["uuid", "name"].forEach(key => {
+                if (_implicitSchema.delete(key)) {
+                    priority.push(key);
+                }
+            });
 
-    if (schema === undefined) {
-        const _implicitSchema = records.map(e => {
-            return new Set(Object.keys(e).filter(key => !key.includes("@")))
-        }).reduce(
-            (acc, current) => new Set([...acc, ...current])
-        );
+            setSchema(priority.concat(Array.from(_implicitSchema)).map(x => {return {label: x, type: "string"}}));
+        }
+    }, []);
 
-        let priority = [];
-        ["uuid", "name"].forEach(key => {
-            if (_implicitSchema.delete(key)) {
-                priority.push(key);
-            }
-        });
-
-        schema = priority.concat(Array.from(_implicitSchema)).map(x => {return {label: x, type: "string"}});
+    const sortTable = (label) => {
+        setRecords([...records.sort((a, b) => a[label] > b[label] ? 1 : -1)]);
     }
 
-    return (
+    return schema ? (
         <StyledTable>
-            <colgroup>
-                {schema.map((item, ii) => <StyledCol class={"data"} />)}
-            </colgroup>
             <thead>
                 <tr>
-                    {schema.map((item, ii) => <StyledHead key={ii} scope={"col"} onClick={evt}>{item.label.toUpperCase()}</StyledHead>)}
+                    {schema.map(({label}, key) => 
+                        <CallbackHeader 
+                            label={label} 
+                            key={key}
+                            onClick={() => sortTable(label)}
+                        />
+                    )}
                 </tr>
             </thead>
             <tbody>
-                {
-                    records.sort((a, b) => a[order] > b[order] ? 1 : -1)
-                        .map((r, i) => <RecordRow schema={schema} record={r} ind={i} />)
-                }
+                {records.map((r, i) => <RecordRow schema={schema} record={r} ind={i}/>)}
             </tbody>
         </StyledTable>
-    )
+    ) : null;
 }
