--- src/demo/index.html
@@ -1,126 +1,0 @@
-<!DOCTYPE html >
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>WebGL and Canvas</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
-</head>
-<link rel="stylesheet" href="../styles/demo.css" type="text/css"/>
-<body>
-
-<script type="module">
-
-    import init, {panic_hook} from "../utils/space.js";
-    import {DataStreamLoop} from "../utils/stream.js";
-    import {Particles, CursorLoop} from "../utils/agent.js";
-    import {RenderingContext, LagrangianParticles} from '../utils/cadet.js';
-    import {TriangularMeshLoop, RectilinearGridLoop, PixelData, HexagonalGrid} from "../utils/tessellate.js";
-
-    (async () => {
-
-        await init();
-        panic_hook();
-
-        [[CursorLoop, {
-            eid: "reticule",
-            context: "2d",
-            caption: "Reticule cursor",
-            font: "12px Arial",
-            autoplay: false
-        }], [RectilinearGridLoop, {
-            eid: "grid",
-            context: "2d",
-            caption: "Rectinlinear grid)",
-            size: [10, 10],
-        }], [TriangularMeshLoop, {
-            eid: "triangle",
-            context: "2d",
-            caption: "Right-triangular irr. network (RTIN)",
-            size: [10, 10],
-        }], [HexagonalGrid, {
-            eid: "hex-grid",
-            context: "2d",
-            caption: "Hexagonal grid",
-            font: "12px Arial",
-        }], [DataStreamLoop, {
-            eid: "data-stream",
-            context: "2d",
-            caption: "DataStream",
-            font: "12px Arial"
-        }], [PixelData, {
-            eid: "pixels",
-            context: "2d",
-            caption: "Draw a pretty picture"
-        }], [LagrangianParticles, {
-            eid: "wind",
-            context: "webgl",
-            source: "../../static/wind.png",
-            metadataFile: "../../static/wind.json",
-            shaders: {
-                draw: ["draw-vertex", "draw-fragment"],
-                screen: ["quad-vertex", "screen-fragment"],
-                update: ["quad-vertex", "update-fragment"],
-                triangle: ["triangle-vertex", "triangle-fragment"],
-            },
-            opacity: 0.996, // how fast the particle trails fade on each frame
-            speed: 0.25, // how fast the particles move
-            drop: 0.003, // how often the particles move to a random place
-            bump: 0.01, // drop rate increase relative to individual particle speed
-            colors: {
-                0.0: '#dd7700',
-                1.0: '#660066'
-            },
-            res: Math.ceil(Math.sqrt(1000))
-        }], [Particles, {
-            eid: "particles",
-            context: "2d",
-            dim: "xyz",
-            padding: 0.0,
-            n: 32
-        }]
-        ].map(target => {
-            let [f, props] = target;
-            let container = document.createElement("div");
-            container.className = "container";
-            let canvas = document.createElement("canvas");
-
-            canvas.width = 200;
-            canvas.height = 200;
-            container.paused = (props.autoplay === undefined) || (!props.autoplay);
-            canvas.id = props.eid;
-            let starting = true;
-            document.body.appendChild(container);
-            props.canvas = canvas;
-            container.appendChild(canvas);
-
-            container.addEventListener("click", () => {
-
-                container.paused = !container.paused;
-
-                if (!container.paused) {
-                    RenderingContext(f, props);
-                    starting = !starting;
-                } else if (!starting) {
-                    container.removeChild(canvas);
-                    canvas = document.createElement("canvas");
-                    canvas.width = 200;
-                    canvas.height = 200;
-                    canvas.id = props.eid;
-                    props.canvas = canvas;
-                    container.appendChild(canvas);
-                    console.log("stop");
-                }
-                container.paused = !container.paused;
-            });
-
-            if (props.autoplay) {
-                RenderingContext(f, props);
-                starting = !starting;
-            }
-
-        })
-    })()
-
-</script>
-</body>
-</html>
