+++ src/pages/404.js
@@ -2,24 +2,18 @@ import React from "react"
 import { graphql } from "gatsby"
 
 import Layout from "../components/Layout"
-import SEO from "../components/seo"
+import SEO from "../components/SEO"
 
-class NotFoundPage extends React.Component {
-  render() {
-    const { data } = this.props
-    const siteTitle = data.site.siteMetadata.title
-
+export default ({location, data: {site: {siteMetadata: {title}}}}) => {
     return (
-      <Layout location={this.props.location} title={siteTitle}>
+      <Layout location={location} title={title}>
         <SEO title="404: Not Found" />
         <h1>Not Found</h1>
         <p>You can't get there from here bub.</p>
       </Layout>
     )
-  }
-}
+};
 
-export default NotFoundPage
 
 export const pageQuery = graphql`
   query {
