--- src/styles/demo.css
@@ -1,277 +1,0 @@
-@charset "UTF-8";
-
-body {
-    margin:0;
-    padding:0;
-}
-
-a {
-    color: #DD66DD;
-}
-
-ul {
-    display: block;
-}
-
-ul li {
-    background: inherit;
-    display: inline-block;
-    font-size: inherit;
-    text-decoration: none;
-    margin: 3px;
-}
-
-body, html{
-    background: #222222;
-    font-size: large;
-    font-family: monospace;
-    color: #cccccc;
-    margin: 0;
-    position: relative;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    overflow-x: hidden;
-    overflow-y: scroll;
-}
-
-
-form {
-    display: inline-block;
-}
-
-select{
-    background: #101010AA;
-    border: solid 1px;
-    display: inline-block;
-    font-size: inherit;
-    color: #ccc;
-    -webkit-appearance: none;  /*Removes default chrome and safari style*/
-    -moz-appearance: none;  /*Removes default style Firefox*/
-}
-
-input, button{
-    background: #101010;
-    border: solid 1px;
-    font-family:inherit;
-    color: #ccc;
-    text-decoration: none;
-}
-
-li:hover{
-    color: #EF5FA1;
-    transition:color 0.5s;
-}
-
-input[type=range] {
-    margin: 10px;
-    width: 50px;
-    background: none;
-}
-
-input[type=range]::-moz-range-track {
-    width: 100%;
-    height: 8px;
-    cursor: pointer;
-    box-shadow: none;
-    background: none;
-    border-radius: 3px;
-    border: 1px solid #FF00FF;
-}
-
-input[type=range]::-moz-range-thumb {
-    box-shadow: none;
-    border: 1px solid #FF00FF;
-    height: 16px;
-    width: 2px;
-    border-radius: 3px;
-    background: #000000;
-}
-
-
-table {
-    table-layout: fixed; /* by headings to make behavior more predictable */
-    width: 100%;
-    border-collapse: collapse; /* remove duplicate borders */
-    border: 3px solid black;
-}
-
-
-th, td, tfoot {
-    border: 1px solid black;
-    padding: 0;
-}
-
-th {
-    padding: 10px;
-}
-
-hr {
-    border-top: 1px solid #ccc;
-}
-
-
-
-tfoot td {
-    position: inherit;
-    padding-right: 5px;
-    text-align: right;
-}
-
-td input {
-    position: relative;
-    width: 95%;
-    height: 100%;
-    padding: 5px;
-    margin: 0;
-    text-align: right;
-    border: none;
-}
-
-
-
-.required:after {
-    content: " R";
-    color: red;
-    font-size: smaller;
-}
-
-.action-success {
-    padding: 0.5em;
-    color: #00d21e;
-    background-color: #eeeeee;
-    border-radius: 0.2em;
-}
-
-.action-failure {
-    padding: 0.5em;
-    color: #ff1408;
-    background-color: #eeeeee;
-    border-radius: 0.2em;
-}
-
-.destructive {
-    background-color: orange;
-}
-
-
-p,
-label {
-    color: #ccc;
-}
-
-canvas {
-    border-radius: 3px;
-    border: 1px solid #FF00FF;
-}
-
-body, html{
-    background: #000000;
-    font-size: large;
-    font-family: monospace;
-    color: #cccccc;
-    margin: 0;
-    position: relative;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    overflow-x: hidden;
-    overflow-y: scroll;
-}
-
-input[type=range] {
-    margin: 10px;
-    width: 50px;
-    background: none;
-}
-
-input[type=range]::-moz-range-track {
-    width: 100%;
-    height: 8px;
-    cursor: pointer;
-    box-shadow: none;
-    background: none;
-    border-radius: 3px;
-    border: 1px solid #FF00FF;
-}
-
-input[type=range]::-moz-range-thumb {
-    box-shadow: none;
-    border: 1px solid #FF00FF;
-    height: 16px;
-    width: 2px;
-    border-radius: 3px;
-    background: #000000;
-}
-
-
-.container {
-    display: inline-flex;
-    flex-direction: row;
-    padding: 5px;
-    outline: none;
-    overflow-wrap: normal;
-    text-align: center;
-}
-
-.container p {
-    display: block;
-    height: 100%;
-    position: absolute;
-    z-index: 10;
-
-}
-
-#bathysphere ul {
-    display: block;
-    padding: 0;
-    margin: 0;
-}
-
-#bathysphere ul li {
-    background: #101010;
-    font-family: inherit;
-    display: inline-block;
-    font-size: inherit;
-    text-decoration: none;
-    border: solid 1px;
-    margin: 3px;
-}
-
-#bathysphere form {
-    display: inline-block;
-}
-
-#bathysphere select{
-    background: #101010AA;
-    border: solid 1px;
-    display: inline-block;
-    font-family:inherit;
-    font-size: inherit;
-    color: #ccc;
-    -webkit-appearance: none;  /*Removes default chrome and safari style*/
-    -moz-appearance: none;  /*Removes default style Firefox*/
-}
-
-#bathysphere input, button{
-    background: #101010;
-    border: solid 1px;
-    font-family:inherit;
-    color: #ccc;
-    text-decoration: none;
-}
-
-#bathysphere li:hover{
-    color: #EF5FA1;
-    transition:color 0.5s;
-}
-
-#bathysphere.container{
-    background: #EF5FA155;
-    position: absolute;
-    margin: 0;
-    padding: 5px;
-    width: 100%;
-    height: auto;
-    z-index: 2;
-}
