+++ src/token-hider.js
@@ -1,11 +1,8 @@
-const axios = require("axios")
 const qs = require("qs")
 
 exports.handler = async function(event, context) {
   // apply our function to the queryStringParameters and assign it to a variable
-  const API_PARAMS = qs.stringify(event.queryStringParameters)
-  // Get env var values defined in our Netlify site UI
-  // TODO: change this
+
   const { API_SECRET = "schnauzer/giant" } = process.env
 
   // TODO: customize your URL
