+++ static/draw-vertex.glsl
