+++ static/rotation-fragment.glsl
